chore: rename common => transition
